added navbar to each page instead of app; added components: BackgroundBanner, Cofounder, photogallery; decoupled about page
