+++ eat-ingrain-client/src/components/Navbar.jsx
@@ -12,4 +12,4 @@ const Navbar = () => {
   )
 }
 
-export default Navbar
+export default Navbar;
